LayoutTests/platform/ios-simulator/css3/shapes/shape-outside/values/shape-outside-shape-arguments-000-expected.txt


PASS 0-valued - inline 
PASS Font relative units - inline 
PASS View relative units - inline 
PASS Fixed units - inline 
PASS Percentage units - inline 
PASS 0-valued - computed 
FAIL Font relative units - computed assert_equals: expected "polygon(16px 7.18px, 8px 16px)" but got "polygon(16px 7.152px, 8px 16px)"(stack: assert@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:705:15
testComputedStyle@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/shapes/shape-outside/values/support/parsing-utils.js:24:22
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:477:40
step@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:1186:30
test@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:429:22
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:475:25
forEach@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2231:30
generate_tests@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:472:16
global code@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/shapes/shape-outside/values/shape-outside-shape-arguments-000.html:58:27)
PASS View relative units - computed 
PASS Fixed units - computed 
PASS Percentage units - computed 
